# specs/001-enterprise-knowledge-rag/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Enterprise Knowledge Intelligence RAG — Backend API
  version: 1.0.0
  summary: Evidence-grounded question answering over publicly accessible documentation.
  description: |
    An independent technical demonstration. Not affiliated with, sponsored by, or endorsed
    by Atlassian or any documentation publisher. Every source returned by this API links back
    to the original publisher page (FR-049, FR-050).

    ## Answering

    `POST /chat/stream` is the primary answering interface: it returns a Server-Sent Events
    stream of pipeline stage events. `POST /chat` returns the same result as a single JSON
    object. Both execute the identical pipeline; the stream is transport, not a different
    code path, so the evaluation runner uses the non-streaming form and measured metrics
    reflect the same work the user sees.

    ## Model reasoning

    No endpoint, response, or event in this API exposes model reasoning or internal
    deliberation (FR-034). Provider calls disable reasoning exposure explicitly; see
    research R-014. The response schemas below are exhaustive and closed
    (`additionalProperties: false`), so a new field cannot appear without a contract change.
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
...
paths:
...
  /documents:
    get:
      tags: [documents]
      summary: Browse indexed documents
      operationId: listDocuments
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: source_id
          in: query
          schema: { type: string, format: uuid }
        - name: state
          in: query
          schema: { $ref: '#/components/schemas/DocumentState' }
        - name: product
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Matching documents
          content:
            application/json:
              schema:
                type: object
                required: [items, total, limit, offset]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Document' }
                  total: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
  /documents/{document_id}:
    parameters:
      - $ref: '#/components/parameters/DocumentIdPath'
    get:
      tags: [documents]
      summary: Retrieve a document
      operationId: getDocument
      responses:
        '200':
          description: The document
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Document' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [documents]
      summary: Delete a document and remove its indexed content
      description: |
        Tombstones the document, then removes its vectors, so the content is unreachable in
        subsequent answers (FR-031, FR-056). Idempotent.
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
  /documents/{document_id}/units:
    get:
      tags: [documents]
      summary: List a document's retrievable units
      description: |
        Returns provenance for each unit — heading path, ordinals, token count, block types —
        but not the unit text, which is held in the vector store. This is the view that lets
        an operator confirm structure-aware chunking without a content dump.
      operationId: listDocumentUnits
      parameters:
        - $ref: '#/components/parameters/DocumentIdPath'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: The document's units in ordinal order
          content:
            application/json:
              schema:
                type: object
                required: [items, total, limit, offset]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/DocumentUnit' }
                  total: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
        '404': { $ref: '#/components/responses/NotFound' }
  /crawl-jobs:
    get:
      tags: [crawl-jobs]
      summary: List crawl jobs
      operationId: listCrawlJobs
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: source_id
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/CrawlJobStatus' }
      responses:
        '200':
          description: Matching crawl jobs
          content:
            application/json:
              schema:
                type: object
                required: [items, total, limit, offset]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/CrawlJob' }
                  total: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
  /crawl-jobs/{job_id}:
    get:
      tags: [crawl-jobs]
      summary: Retrieve a crawl job
      operationId: getCrawlJob
      parameters:
        - $ref: '#/components/parameters/JobIdPath'
      responses:
        '200':
          description: The crawl job
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CrawlJob' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  parameters:
...
    DocumentIdPath:
      name: document_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    JobIdPath:
      name: job_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
...
  responses:
...
    NotFound:
      description: No such resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, request_id]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - SSRF_BLOCKED
                - INVALID_URL
                - ROBOTS_DISALLOWED
                - UNAUTHORIZED
                - NOT_FOUND
                - CONFLICT
                - CRAWL_ALREADY_RUNNING
                - DOCUMENT_DELETED
                - RATE_LIMITED
                - PROVIDER_ERROR
                - PROVIDER_RATE_LIMITED
                - VECTOR_SERVICE_UNAVAILABLE
                - INTERNAL_ERROR
            message: { type: string }
            request_id: { type: string, format: uuid }
            details:
              type: object
              additionalProperties: true
              description: Field names and enumerated values only. Never free-form user input.
...
    DocumentState:
      type: string
      enum: [discovered, crawling, processed, indexed, failed, deleted]
    Document:
      type: object
      additionalProperties: false
      required: [id, source_id, url, state, vectors_live, unit_count]
      properties:
        id: { type: string, format: uuid }
        source_id: { type: string, format: uuid }
        url: { type: string, format: uri }
        canonical_url: { type: [string, 'null'], format: uri }
        title: { type: [string, 'null'] }
        product: { type: [string, 'null'] }
        category: { type: [string, 'null'] }
        page_type: { type: [string, 'null'] }
        language: { type: [string, 'null'] }
        state: { $ref: '#/components/schemas/DocumentState' }
        state_detail: { type: [string, 'null'] }
        vectors_live:
          type: boolean
          description: |
            True when vectors for this document are retrievable. Stays true for a document in
            `failed` that was previously indexed, because a failed re-crawl leaves prior content
            live (FR-054). This is why it is not derivable from `state`.
        unit_count: { type: integer }
        content_fingerprint: { type: [string, 'null'] }
        source_modified_at: { type: [string, 'null'], format: date-time }
        last_crawled_at: { type: [string, 'null'], format: date-time }
        indexed_at: { type: [string, 'null'], format: date-time }
        tombstoned_at: { type: [string, 'null'], format: date-time }
        created_at: { type: string, format: date-time }
    DocumentUnit:
      type: object
      additionalProperties: false
      description: |
        Provenance only. Unit text is held in the vector store and is deliberately not returned
        by this endpoint, so a corpus view cannot become a content dump.
      required: [id, document_id, ordinal, heading_path, token_count, block_types]
      properties:
        id: { type: string, format: uuid }
        document_id: { type: string, format: uuid }
        ordinal: { type: integer, minimum: 0 }
        vector_id: { type: string }
        heading_path: { type: array, items: { type: string } }
        block_types: { type: array, items: { type: string } }
        token_count: { type: integer, minimum: 1 }
        overlap_tokens: { type: integer, minimum: 0 }
        text_fingerprint: { type: string }
    CrawlJobStatus:
      type: string
      enum: [queued, running, completed, completed_with_errors, failed, cancelled]
    CrawlJob:
      type: object
      additionalProperties: false
      required: [id, source_id, target_url, status, requested_at, counters]
      properties:
        id: { type: string, format: uuid }
        source_id: { type: string, format: uuid }
        target_url: { type: string, format: uri }
        status: { $ref: '#/components/schemas/CrawlJobStatus' }
        requested_at: { type: string, format: date-time }
        started_at: { type: [string, 'null'], format: date-time }
        finished_at: { type: [string, 'null'], format: date-time }
        counters:
          type: object
          additionalProperties: false
          required: [discovered, processed, unchanged, skipped, failed]
          properties:
            discovered: { type: integer }
            processed: { type: integer }
            unchanged: { type: integer, description: Pages whose fingerprint matched; not re-indexed (SC-012). }
            skipped: { type: integer }
            failed: { type: integer }
        skipped_reasons: { type: object, additionalProperties: { type: integer } }
        error_code: { type: [string, 'null'] }
        error_message: { type: [string, 'null'] }
...